+++ src/vue/modules/create-asset-form/components/information-step-form.vue
@@ -40,7 +40,7 @@
         <input-field
           white-autofill
           type="number"
-          :min="MIN_AMOUNT"
+          :min="0"
           :max="MAX_AMOUNT"
           :step="MIN_AMOUNT"
           v-model="form.maxIssuanceAmount"