+++ src/vue/modules/issuance-form/components/create-issuance-form.vue
@@ -31,7 +31,7 @@
           <input-field
             white-autofill
             type="number"
-            :min="MIN_AMOUNT"
+            :min="0"
             :max="form.asset.availableForIssuance"
             :step="MIN_AMOUNT"
             v-model="form.amount"