+++ src/vue/modules/loyalty-points/create-invoice-form/components/invoice-form.vue
@@ -28,7 +28,7 @@
           :label="'create-invoice-form.amount-lbl' | globalize({
             asset: form.asset
           })"
-          :min="MIN_AMOUNT"
+          :min="0"
           :max="MAX_AMOUNT"
           :step="MIN_AMOUNT"
           :error-message="getFieldErrorMessage(