+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Changed
+- Now loading private document URL in file field
+
+#### Fixed
+- Disabled state for amount input field
+- Minimum & maximum normalization for input field
+- Displaying balance on assets page
+- Displaying no data row on fees table
+- Displaying "My sales" tab for non-corporate users
+- Selecting asset on deposit form 
+
 ## [1.8.0-x.2] - 2019-06-05
 #### Fixed
 - Bug with investing in sale