+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
+#### Added
+- Asset code validator
+
 #### Changed
 - Now loading private document URL in file field