+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Added
+- Asset code validator
+
+#### Changed
+- Now loading private document URL in file field
+
+#### Fixed
+- Disabled state for amount input field
+- Minimum & maximum normalization for input field
+- Displaying balance on assets page
+- Displaying no data row on fees table
+- Displaying "My sales" tab for non-corporate users
+- Selecting asset on deposit form 
+
+## [1.8.0-x.3] - 2019-06-05
+
+## [1.8.0-x.2] - 2019-06-05
+#### Fixed
+- Bug with investing in sale
+
 ## [1.8.0-x.1] - 2019-06-05
 #### Added
 - Skeleton loading
@@ -663,7 +684,9 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.8.0-x.1...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.8.0-x.3...HEAD
+[1.8.0-x.3]: https://github.com/tokend/web-client/compare/1.8.0-x.2...1.8.0-x.3
+[1.8.0-x.2]: https://github.com/tokend/web-client/compare/1.8.0-x.1...1.8.0-x.2
 [1.8.0-x.1]: https://github.com/tokend/web-client/compare/1.8.0-x.0...1.8.0-x.1
 [1.8.0-x.0]: https://github.com/tokend/web-client/compare/1.7.0...1.8.0-x.0
 [1.7.0]: https://github.com/tokend/web-client/compare/1.7.0-rc.4...1.7.0